requirements.txt
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
pygments==2.19.2
